requirements.txt
django==4.2.4
djangorestframework==3.14.0
openai==0.27.10
pytube==15.0.0
pydub==0.25.1
python-dotenv==1.0.0